backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pytest==8.3.2
supabase==2.7.4
python-multipart==0.0.9
watchdog==4.0.1
openai==1.51.2
boto3==1.34.162
certifi==2024.8.30
